Enable build and run for Gradle, put Task subtypes in their own files
