+++ data/list.txt
@@ -1,4 +1,4 @@
-list: 7
+list: 8
 T : 1 : eat
 E : 0 : sleep : 30/12/2025 22:00 : 31/12/2025 06:00
 D : 1 : shit : 03/02/2025 13:00
@@ -6,3 +6,4 @@ T : 1 : eat more
 D : 0 : 2103 iP : 31/01/3099 23:59
 T : 0 : dinner
 T : 1 : supper
+T : 0 : hello world
