+++ src/main/java/luke/command/Command.java
@@ -1,0 +1,7 @@
+package luke.command;
+
+import luke.component.*;
+
+public abstract class Command {
+    public abstract void execute(TaskList tl, Ui ui, Storage st);
+}
